requirements.txt
asgiref==3.8.1
Django==5.1.2
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
kafka-python==2.0.2
confluent-kafka==2.6.0
six==1.16.0
